+++ models/media_viewing.js
@@ -1,0 +1,13 @@
+const { Sequelize, DataTypes } = require("sequelize");
+const sequelize = require("../database/database.js");
+
+const Media_viewing = sequelize.define("media_viewing", {
+    mediaId: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+    },
+    userId: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+    },
+  });
